+++ src/app/utils/login.py
@@ -11,8 +11,8 @@ from logfire import configure as logfire_conf
 from wandb import login as wandb_login
 from weave import init as weave_init
 
-from app.agents.llm_model_funs import get_api_key
 from app.data_models.app_models import AppEnv
+from app.llms.providers import get_api_key
 from app.utils.error_messages import generic_exception
 from app.utils.log import logger